requirement.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1733730548347/work
astunparse @ file:///home/conda/feedstock_root/build_artifacts/astunparse_1736248061654/work
audioread==3.0.1
blinker==1.9.0
Brotli @ file:///croot/brotli-split_1736182456865/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1739515848642/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1735929714516/work
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
Flask==3.1.0
flatbuffers @ file:///home/conda/feedstock_root/build_artifacts/python-flatbuffers_1739279199749/work
fonttools==4.56.0
gast @ file:///home/conda/feedstock_root/build_artifacts/gast_1733299680490/work
google-pasta @ file:///home/conda/feedstock_root/build_artifacts/google-pasta_1733852499742/work
grpcio @ file:///croot/grpc-split_1716834572340/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
h5py @ file:///croot/h5py_1737051113867/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
keras @ file:///home/conda/feedstock_root/build_artifacts/keras_1736538740036/work
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1710435156458/work
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1733250460757/work
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
ml_dtypes @ file:///croot/ml_dtypes_1722369114743/work
msgpack==1.1.0
namex @ file:///home/conda/feedstock_root/build_artifacts/namex_1733858070416/work
numba==0.61.0
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp311-cp311-linux_x86_64.whl#sha256=2f550fcfe63acca79d636ba578e32b7ec2999e0f9cd88120af57d1497ad15890
opencv-python==4.10.0
opencv-python-headless==4.10.0
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
optree @ file:///croot/optree_1723069484006/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.3
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1736243443484/work
pyparsing==3.2.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1733217035951/work
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1733342254348/work/dist
scikit-learn==1.6.1
scipy==1.15.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1733380938961/work
soundfile==0.13.1
soxr==0.5.0.post1
tensorboard @ file:///home/conda/feedstock_root/build_artifacts/tensorboard_1723713090305/work/tensorboard-2.17.1-py3-none-any.whl#sha256=253701a224000eeca01eee6f7e978aea7b408f60b91eb0babdb04e78947b773e
tensorboard_data_server @ file:///croot/tensorboard-data-server_1724172556187/work/tensorboard_data_server-0.7.0-py3-none-manylinux2014_x86_64.whl#sha256=c166a9f297fa12121b89b94be4a4cc150aa7996c9598410ba057ecd592eea8c7
tensorflow @ file:///opt/conda/conda-bld/tensorflow-split_1729547219394/work/tensorflow_pkg/tensorflow-2.17.0-cp311-cp311-linux_x86_64.whl#sha256=1f15e1d189a071ddeb5c9f311f831ca939e502411ddc102b9efc257686466792
tensorflow_estimator @ file:///opt/conda/conda-bld/tensorflow-split_1729547219394/work/tensorflow-estimator/wheel_dir/tensorflow_estimator-2.15.0-py2.py3-none-any.whl#sha256=4fc96f61b1762702b8bce247493e622bdd079448df532778661223a7c780afd3
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1733754631889/work
threadpoolctl==3.5.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1733188668063/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1734859416348/work
Werkzeug==3.1.3
wrapt @ file:///croot/wrapt_1736540904746/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zstandard @ file:///croot/zstandard_1731356346222/work
